refactor: move acceptance-test into bootstrap-blocking
