+++ bootstrap-blocking/pom.xml
@@ -3,19 +3,16 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <groupId>packagename</groupId>
     <artifactId>artifactName</artifactId>
+    <groupId>packagename</groupId>
     <version>1.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>acceptance-test</artifactId>
+  <artifactId>bootstrap-blocking</artifactId>
+  <properties>
+    <start-class>packagename.boot.blocking.ExampleApplication</start-class>
+  </properties>
   <dependencies>
-    <dependency>
-      <groupId>packagename</groupId>
-      <artifactId>bootstrap</artifactId>
-    </dependency>
-    <!-- Important: DO NOT DELETE the below dependencies though might seem like redundant to that of
-    bootstrap but they are required for jacoco and pitest report aggregation -->
     <!-- left adapter -->
     <dependency>
       <groupId>packagename</groupId>
@@ -31,15 +28,25 @@
       <groupId>packagename</groupId>
       <artifactId>jpa-adapter</artifactId>
     </dependency>
+    <!-- Frameworks & Libraries -->
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-actuator</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-devtools</artifactId>
+    </dependency>
+    <!-- temporary database -->
+    <dependency>
+      <groupId>com.h2database</groupId>
+      <artifactId>h2</artifactId>
+      <scope>runtime</scope>
+    </dependency>
     <!-- Test -->
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-test</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.h2database</groupId>
-      <artifactId>h2</artifactId>
       <scope>test</scope>
     </dependency>
     <!-- Cucumber dependencies -->
@@ -85,16 +92,6 @@
         </executions>
       </plugin>
       <plugin>
-        <groupId>org.jetbrains.kotlin</groupId>
-        <artifactId>kotlin-maven-plugin</artifactId>
-        <configuration>
-          <compilerPlugins>
-            <plugin>jpa</plugin>
-            <plugin>spring</plugin>
-          </compilerPlugins>
-        </configuration>
-      </plugin>
-      <plugin>
         <groupId>org.jacoco</groupId>
         <artifactId>jacoco-maven-plugin</artifactId>
         <executions>
@@ -108,11 +105,44 @@
         </executions>
       </plugin>
       <plugin>
+        <groupId>org.jetbrains.kotlin</groupId>
+        <artifactId>kotlin-maven-plugin</artifactId>
+        <configuration>
+          <compilerPlugins>
+            <plugin>jpa</plugin>
+            <plugin>spring</plugin>
+          </compilerPlugins>
+        </configuration>
+      </plugin>
+      <plugin>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-maven-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>repackage</id>
+            <configuration>
+              <classifier>exec</classifier>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <groupId>com.google.cloud.tools</groupId>
+        <artifactId>jib-maven-plugin</artifactId>
+      </plugin>
+      <plugin>
         <groupId>com.societegenerale.commons</groupId>
         <artifactId>arch-unit-maven-plugin</artifactId>
         <configuration>
           <rules>
             <preConfiguredRules>
+              <rule>com.societegenerale.commons.plugin.rules.NoStandardStreamRuleTest</rule>
+              <rule>com.societegenerale.commons.plugin.rules.NoJodaTimeRuleTest</rule>
+              <rule>com.societegenerale.commons.plugin.rules.NoJavaUtilDateRuleTest</rule>
+              <rule>com.societegenerale.commons.plugin.rules.NoPrefixForInterfacesRuleTest</rule>
+              <rule>com.societegenerale.commons.plugin.rules.NoPublicFieldRuleTest</rule>
+              <rule>com.societegenerale.commons.plugin.rules.NoInjectedFieldTest</rule>
+              <rule>com.societegenerale.commons.plugin.rules.NoAutowiredFieldTest</rule>
               <rule>com.societegenerale.commons.plugin.rules.NoJunitAssertRuleTest</rule>
               <rule>com.societegenerale.commons.plugin.rules.NoPowerMockRuleTest</rule>
               <rule>com.societegenerale.commons.plugin.rules.NoTestIgnoreRuleTest</rule>
@@ -153,3 +183,4 @@
     </profile>
   </profiles>
 </project>
+
