+++ bootstrap-blocking/src/main/kotlin/packagename/boot/blocking/config/BlockingBootstrapConfig.kt
@@ -1,4 +1,4 @@
-package packagename.boot.config
+package packagename.boot.blocking.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -10,9 +10,9 @@ import packagename.repository.config.JpaAdapterConfig
 
 @Configuration
 @Import(JpaAdapterConfig::class)
-class BootstrapConfig {
+class BlockingBootstrapConfig {
 
-  @Bean
+  @Bean("blocking-request-example")
   fun getRequestExample(obtainExample: ObtainExample): RequestExample {
     return ExampleDomain(obtainExample)
   }
