+++ bootstrap-blocking/src/test/kotlin/packagename/boot/blocking/test/cucumber/ExampleStepDef.kt
@@ -1,4 +1,4 @@
-package packagename.cucumber
+package packagename.boot.blocking.test.cucumber
 
 import io.cucumber.datatable.DataTable
 import io.cucumber.java8.En
@@ -8,15 +8,16 @@ import org.springframework.boot.web.server.LocalServerPort
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import packagename.domain.model.Example
-import packagename.rest.representation.ExampleInfo
 import packagename.repository.dao.ExampleDao
 import packagename.repository.entity.ExampleEntity
 import packagename.rest.exception.ExampleExceptionResponse
+import packagename.rest.representation.ExampleInfo
 
 
 class ExampleStepDef(restTemplate: TestRestTemplate, exampleDao: ExampleDao) : En {
 
   companion object {
+
     private const val LOCALHOST = "http://localhost:"
     private const val API_URI = "/api/v1/examples"
   }
@@ -27,14 +28,23 @@ class ExampleStepDef(restTemplate: TestRestTemplate, exampleDao: ExampleDao) : E
 
   init {
 
-    DataTableType { row: Map<String, String> -> Example(row["code"].toString().toLong(), row["description"].toString()) }
-    DataTableType { row: Map<String, String> -> ExampleEntity(code = row["code"].toString().toLong(), description = row["description"].toString()) }
+    DataTableType { row: Map<String, String> ->
+      Example(
+        row["code"].toString().toLong(),
+        row["description"].toString()
+      )
+    }
+    DataTableType { row: Map<String, String> ->
+      ExampleEntity(
+        code = row["code"].toString().toLong(), description = row["description"].toString()
+      )
+    }
 
     Before { _ -> exampleDao.deleteAll() }
     After { _ -> exampleDao.deleteAll() }
 
     Given("the following examples exists in the library") { dataTable: DataTable ->
-      val examples = dataTable.asList<ExampleEntity>(ExampleEntity::class.java)
+      val examples = dataTable.asList(ExampleEntity::class.java)
       exampleDao.saveAll(examples)
     }
 
@@ -63,13 +73,16 @@ class ExampleStepDef(restTemplate: TestRestTemplate, exampleDao: ExampleDao) : E
     }
 
     Then("the user gets the following examples") { dataTable: DataTable ->
-      val expectedExamples = dataTable.asList<Example>(Example::class.java)
+      val expectedExamples = dataTable.asList(Example::class.java)
       assertThat(responseEntity.statusCode).isEqualTo(HttpStatus.OK)
       val body = responseEntity.body
       assertThat(body).isNotNull
       when (body) {
-        is ExampleInfo -> assertThat(body.examples).isNotEmpty.extracting("description").containsAll(expectedExamples.map { it.description })
-        is Example -> assertThat(body).isNotNull.extracting("description").isEqualTo(expectedExamples.first().description)
+        is ExampleInfo -> assertThat(body.examples).isNotEmpty.extracting("description")
+          .containsAll(expectedExamples.map { it.description })
+
+        is Example -> assertThat(body).isNotNull.extracting("description")
+          .isEqualTo(expectedExamples.first().description)
       }
     }
   }
