+++ bootstrap-blocking/src/test/kotlin/packagename/boot/blocking/test/cucumber/SpringCucumberTestConfig.kt
@@ -1,11 +1,14 @@
-package packagename.cucumber
+package packagename.boot.blocking.test.cucumber
 
 import io.cucumber.spring.CucumberContextConfiguration
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
-import packagename.boot.ExampleApplication
+import packagename.boot.blocking.ExampleApplication
 
-@SpringBootTest(classes = [ExampleApplication::class], webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@SpringBootTest(
+  classes = [ExampleApplication::class],
+  webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
+)
 @CucumberContextConfiguration
 @ActiveProfiles("test")
 class SpringCucumberTestConfig
